requirements.txt
click>=8.0
rich>=13.0
python-dotenv>=1.0
anthropic>=0.18
Pillow>=10.0
openai>=1.0
